+++ pages/admin-pending.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import AdminPending from '../src/app/components/AdminPendingStudies.js';
+import AdminPending from '../src/app/components/AdminPendingApproval/AdminPendingStudies.js';
 
 const AdminPendingPage = () => {
   return (
