+++ pages/pending.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { db } from '../src/app/firebase';
-import { collection, getDocs, updateDoc, doc } from 'firebase/firestore';
-import Modal from '../src/app/components/ResearcherFeedbackModal';
+import { collection, query, where, getDocs } from 'firebase/firestore';
+import Modal from '../src/app/components/ResearcherPendingStudies/ResearcherFeedbackPopUp';
 
 const Pending = () => {
   const [submissions, setSubmissions] = useState([]);
@@ -13,23 +13,14 @@ const Pending = () => {
   useEffect(() => {
     const fetchSubmissions = async () => {
       try {
-        const querySnapshot = await getDocs(collection(db, "researchStudies"));
+        // Adjusted query to fetch submissions with 'researcherPending' status
+        const q = query(collection(db, "researchStudies"), where("status", "==", "researcherPending"));
+        const querySnapshot = await getDocs(q);
         const submissionsData = querySnapshot.docs.map(doc => ({
           id: doc.id,
           ...doc.data()
         }));
-
-        // Fetch feedback for each submission
-        const feedbackPromises = submissionsData.map(submission => {
-          return getDocs(collection(db, `feedback/${submission.id}/comments`))
-            .then(querySnapshot => {
-              const comments = querySnapshot.docs.map(doc => doc.data());
-              return { ...submission, feedback: comments };
-            });
-        });
-
-        const submissionsWithFeedback = await Promise.all(feedbackPromises);
-        setSubmissions(submissionsWithFeedback.filter(submission => submission.status === 'pending'));
+        setSubmissions(submissionsData);
       } catch (err) {
         setError('Failed to fetch submissions. Please try again.');
       } finally {
@@ -39,34 +30,6 @@ const Pending = () => {
 
     fetchSubmissions();
   }, []);
-
-  const handleApprove = async (id) => {
-    try {
-      const submissionRef = doc(db, "researchStudies", id);
-      await updateDoc(submissionRef, { status: 'approved' });
-      setSubmissions(submissions.filter(submission => submission.id !== id));
-    } catch (err) {
-      setError('Failed to update submission status. Please try again.');
-    }
-  };
-
-  const handleRevise = async (id) => {
-    try {
-      // Your logic to handle revision
-      console.log('Revision logic here');
-    } catch (err) {
-      setError('Failed to handle revision. Please try again.');
-    }
-  };
-
-  const handleResubmit = async (id) => {
-    try {
-      // Your logic to handle resubmission
-      console.log('Resubmit logic here');
-    } catch (err) {
-      setError('Failed to handle resubmission. Please try again.');
-    }
-  };
 
   const handleViewFeedback = (submission) => {
     setCurrentSubmission(submission);
@@ -85,10 +48,7 @@ const Pending = () => {
               <li key={submission.id} style={{ display: 'flex', justifyContent: 'space-between', marginBottom: '10px', alignItems: 'center', padding: '10px', border: '1px solid #ccc', borderRadius: '5px', background: '#f9f9f9' }}>
                 <button onClick={() => handleViewFeedback(submission)} style={{ marginRight: '10px', background: '#007bff', color: 'white', border: 'none', borderRadius: '5px', padding: '5px 10px', cursor: 'pointer' }}>View Feedback</button>
                 <span style={{ flexGrow: 1 }}>{submission.title}</span>
-                <div>
-                  <button onClick={() => handleRevise(submission.id)} style={{ marginRight: '5px', background: '#ffc107', color: 'white', border: 'none', borderRadius: '5px', padding: '5px 10px', cursor: 'pointer' }}>Revise</button>
-                  <button onClick={() => handleResubmit(submission.id)} style={{ background: '#007bff', color: 'white', border: 'none', borderRadius: '5px', padding: '5px 10px', cursor: 'pointer' }}>Resubmit</button>
-                </div>
+                {/* Implement logic for resubmitting submissions */}
               </li>
             ))}
           </ul>
