+++ src/app/services/firestoreOperations.js
@@ -1,16 +1,15 @@
 import { db } from "../firebase";
-import { doc, getDocs, setDoc } from "firebase/firestore";
-import { updateDoc } from "firebase/firestore";
+import { doc, setDoc, updateDoc } from "firebase/firestore";
 
 export const createUserProfileDocument = async (userAuth, additionalData) => {
   if (!userAuth) return;
 
   const userRef = doc(db, `users/${userAuth.uid}`);
-
   await setDoc(userRef, {
     uid: userAuth.uid,
     email: userAuth.email,
-    role: "researcher", // Default role,
+    role: "researcher", // Default role
+    ...additionalData, // Pass additional data to the user profile document
   });
 };
 
@@ -19,4 +18,17 @@ export const updateUserRole = async (uid, newRole) => {
   await updateDoc(userRef, {
     role: newRole,
   });
+};
+
+export const updateResearchProposalStatus = async (proposalId, newStatus, feedback = '') => {
+  const proposalRef = doc(db, `researchStudies`, proposalId);
+  try {
+    await updateDoc(proposalRef, {
+      status: newStatus,
+      adminFeedback: feedback // adminFeedback field is updated along with status
+    });
+  } catch (error) {
+    console.error('Error updating research study status: ', error);
+    throw new Error("Could not update research study status."); // Throw an error to handle if needed
+  }
 };
